requirements.txt
flask
motor
aiohttp
python-dotenv
websockets
